requirements.txt
fastapi~=0.115.8
ffmpeg~=1.4
python-multipart~=0.0.20
openai-whisper~=20240930
setuptools-rust~=1.10.2
pytesseract~=0.3.13
redis~=5.2.1
python-dotenv~=1.0.1